backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.25.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=1.0.0
pydantic>=2.0.0
